fixed events component: still need to tweek render/data grabbed
