+++ frontend/src/components/Splash/index.js
@@ -1,28 +1,11 @@
-import { NavLink, useParams } from 'react-router-dom'
-import { useSelector, useDispatch } from 'react-redux'
-import { useState, useEffect } from 'react'
+import { NavLink } from 'react-router-dom'
+import { useSelector } from 'react-redux'
+import EventsList from '../EventsList'
 import "./SplashPage.css"
 
-import { fetchEvents } from '../../store/event';
 
 const SplashPage = () => {
   const sessionUser = useSelector(state => state.session.user);
-
-  const dispatch = useDispatch();
-  const { eventId } = useParams();
-
-  const events = useSelector(state => {
-    return state.events.list.map(eventId => state.events[eventId]);
-  });
-
-  useEffect(() => {
-    dispatch(fetchEvents());
-  }, [dispatch]);
-
-  if (!events) {
-    return null;
-  }
-
 
 
   if(sessionUser) {
@@ -34,29 +17,10 @@ const SplashPage = () => {
               <p className='title'> Welcome to Dive Events! </p>
               <img className='tank-photo' src="/images/scuba-background.jpg" alt=''/>
               <p className='info-p'> Please look below for upcoming diving events! </p>
-              <div className='events-links'>
-                {events.map(event => {
-                  return (
-                    <NavLink key={event.name} to={`/events/${event.id}`}>
-                      <div
-                        className={
-                          Number.parseInt(eventId) === event.id
-                            ? 'nav-entry is-selected'
-                            : 'nav-entry'
-                        }
-                      >
-                        <div>
-                          <div className='event-name'>{event.name}</div>
-                          <div className='event-date'>{event.date}</div>
-                        </div>
-                      </div>
-                    </NavLink>
-                  )
-                })}
-
-              </div>
             </div>
-              <img className='beach-background' src="/images/beach-background.jpg" alt=''/>
+            <div>
+              <EventsList />
+            </div>
         </div>
 
       </>
