+++ frontend/src/store/event.js
@@ -46,36 +46,32 @@ const deleteEvent = deletedEvent => {
 
 
 //thunk action creators
-export const fetchEvents = events => async dispatch => {
+export const fetchEvents = () => async dispatch => {
   const res = await fetch(`/api/events`);
 
   if (res.ok) {
     const events = await res.json();
     dispatch(getEvents(events));
-    return events;
   }
 }
 
-
-
-
 //reducer
-const initialState = {};
+
+const initialState = {
+  events: [],
+};
 
 const eventsReducer = (state = initialState, action) => {
   switch (action.type) {
 
     case GET_EVENTS: {
-      state.events = action.events;
-      return {...state};
+      const newState = {};
+      action.events.forEach((event) => (newState[event.id] = event));
+      return newState;
     }
 
-
-
-
-
-
-
+    default:
+      return state;
   }
 }
